requirements.txt
asgiref==3.4.1
Django==3.2.6
django-environ==0.4.5
django-taggit==1.2.0
Markdown==3.2.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
python-dotenv==0.19.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==4.0.1
